[font] cleanup comments
