+++ font/opentype/reader.go
@@ -41,9 +41,6 @@ var (
 // practice, it seems to work.
 const dfontResourceDataOffset = 0x00000100
 
-// Resource is a combination of io.Reader, io.Seeker and io.ReaderAt.
-// This interface is satisfied by most things that you'd want
-// to parse, for example *os.File, io.SectionReader or *bytes.Reader.
 type Resource interface {
 	Read([]byte) (int, error)
 	ReadAt([]byte, int64) (int, error)
